BibRank: wrong diffs when checking citation losses
* The differ was using the cites dictionary twice when checking for
  the overall citation balances in bibrank -w citation.
